Refactor: Move partydonation tests to its own file
